services/transcription-service/requirements.txt
flask
gunicorn
google-cloud-speech
google-cloud-storage
google-cloud-pubsub
google-cloud-firestore
pydub